+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to similar are documented here.
 ## 0.3.0
 
 * Added grapheme and character level diffing utilities.
+* `DiffOp::as_tag_tuple` is now taking the argument by reference.
 
 ## 0.2.0